+++ src/transitive/infallible/from.rs
@@ -1,34 +1,41 @@
 use std::iter::once;
 
-use darling::{util::PathList, FromAttributes};
 use proc_macro2::TokenStream;
 use quote::{quote, ToTokens};
+use syn::{
+    parse::{Parse, ParseStream},
+    punctuated::Punctuated,
+    Path, Result as SynResult, Token,
+};
 
-use crate::transitive::attr::AttrWithIdent;
+use crate::transitive::TokenizablePath;
 
-#[derive(FromAttributes)]
-#[darling(attributes(transitive))]
-pub struct TransitiveFrom {
-    from: PathList,
+/// Path corresponding to a [`#[transitive(from(..))`] path.
+pub struct TransitionFrom(Punctuated<Path, Token![,]>);
+
+impl Parse for TransitionFrom {
+    fn parse(input: ParseStream) -> SynResult<Self> {
+        Punctuated::parse_terminated(input).map(Self)
+    }
 }
 
-impl ToTokens for AttrWithIdent<'_, &TransitiveFrom> {
+impl ToTokens for TokenizablePath<'_, &TransitionFrom> {
     fn to_tokens(&self, tokens: &mut TokenStream) {
         let name = self.ident;
-        let first = self.data.from.first();
+        let (impl_generics, ty_generics, where_clause) = self.generics.split_for_impl();
+        let first = self.path.0.first();
 
         let stmts = self
-            .data
-            .from
+            .path
+            .0
             .iter()
-            .map(|ty| quote! {let val = #ty::from(val);})
-            .chain(once(quote! {let val = #name::from(val);}));
+            .map(|ty| quote! {let val: #ty = core::convert::From::from(val);})
+            .chain(once(quote! {core::convert::From::from(val)}));
 
         let expanded = quote! {
-            impl core::convert::From<#first> for #name {
-                fn from(val: #first) -> #name {
+            impl #impl_generics core::convert::From<#first> for #name #ty_generics #where_clause {
+                fn from(val: #first) -> Self {
                     #(#stmts)*
-                    val
                 }
             }
         };