+++ src/transitive/infallible/mod.rs
@@ -1,5 +1,5 @@
 mod from;
 mod into;
 
-pub use from::TransitiveFrom;
-pub use into::TransitiveInto;
+pub use from::TransitionFrom;
+pub use into::TransitionInto;